Initial commit of MVP java Wordle program.
A simple java terminal program that runs Wordle locally in a terminal.

You will get 5 tried to guess a 5 letter word from the dictionary. Anytime you guess the correct letter and position it will be indicated as such.

You can see what is the word by typing "showmeword" or exit to terminate the program.

Co-Authored-By: Zain Aziz <14876833+zanyaziz@users.noreply.github.com>